+++ spec/believer/extensions/will_paginate_spec.rb
@@ -1,0 +1,12 @@
+require 'spec_helper'
+
+describe 'Believer::Extensions::WillPaginate' do
+
+  before :each do
+    @count = 10
+  end
+
+  it '' do
+
+  end
+end