+++ spec/support/test_classes.rb
@@ -65,7 +65,7 @@ module Test
           :host => '127.0.0.1',
           :keyspace => 'believer_test_space',
           :believer => {
-              :logger => {:level => ::Logger::DEBUG}
+              :logger => {:level => ::Logger::INFO}
           }
       }
     end